+++ packages/core/src/common/multi.c
@@ -1,5 +1,8 @@
 #include <combo.h>
 #include <combo/net.h>
+#include <combo/player.h>
+#include <combo/config.h>
+#include <combo/item.h>
 
 u32 gMultiMarkChests;
 u32 gMultiMarkCollectibles;
@@ -277,9 +280,9 @@ static void setXflagsMarkMm(GameState_Play* play, int sliceId, int sceneId, int 
     comboXflagsSetMm(&xf);
 }
 
-void multiSetMarkedOot(GameState_Play* play, u8 ovType, u8 sceneId, u8 roomId, u8 id)
+void Multi_SetMarkedOot(GameState_Play* play, u8 ovType, u8 sceneId, u8 roomId, u8 id)
 {
-    if (!comboConfig(CFG_MULTIPLAYER))
+    if (!Config_Flag(CFG_MULTIPLAYER))
         return;
 
     switch (ovType)
@@ -321,9 +324,9 @@ void multiSetMarkedOot(GameState_Play* play, u8 ovType, u8 sceneId, u8 roomId, u
     }
 }
 
-void multiSetMarkedMm(GameState_Play* play, u8 ovType, u8 sceneId, u8 roomId, u8 id)
+void Multi_SetMarkedMm(GameState_Play* play, u8 ovType, u8 sceneId, u8 roomId, u8 id)
 {
-    if (!comboConfig(CFG_MULTIPLAYER))
+    if (!Config_Flag(CFG_MULTIPLAYER))
         return;
 
     switch (ovType)
@@ -362,9 +365,9 @@ void multiSetMarkedMm(GameState_Play* play, u8 ovType, u8 sceneId, u8 roomId, u8
     }
 }
 
-int multiIsMarkedOot(GameState_Play* play, u8 ovType, u8 sceneId, u8 roomId, u8 id)
+int Multi_IsMarkedOot(GameState_Play* play, u8 ovType, u8 sceneId, u8 roomId, u8 id)
 {
-    if (!comboConfig(CFG_MULTIPLAYER))
+    if (!Config_Flag(CFG_MULTIPLAYER))
         return 0;
 
     switch (ovType)
@@ -398,9 +401,9 @@ int multiIsMarkedOot(GameState_Play* play, u8 ovType, u8 sceneId, u8 roomId, u8 
     return 0;
 }
 
-int multiIsMarkedMm(GameState_Play* play, u8 ovType, u8 sceneId, u8 roomId, u8 id)
+int Multi_IsMarkedMm(GameState_Play* play, u8 ovType, u8 sceneId, u8 roomId, u8 id)
 {
-    if (!comboConfig(CFG_MULTIPLAYER))
+    if (!Config_Flag(CFG_MULTIPLAYER))
         return 0;
 
     switch (ovType)
@@ -570,11 +573,11 @@ static void processMessages(GameState_Play* play, NetContext* net)
     }
 }
 
-void comboMultiProcessMessages(GameState_Play* play)
+void Multi_ProcessMessages(GameState_Play* play)
 {
     NetContext* ctx;
 
-    if (!comboConfig(CFG_MULTIPLAYER))
+    if (!Config_Flag(CFG_MULTIPLAYER))
         return;
 
     ctx = netMutexLock();
@@ -597,9 +600,9 @@ static void drawSingleWisp(GameState_Play* play, const PlayerWisp* wisp)
     CLOSE_DISPS();
 }
 
-void comboMultiDrawWisps(GameState_Play* play)
+void Multi_DrawWisps(GameState_Play* play)
 {
-    if (!comboConfig(CFG_MULTIPLAYER))
+    if (!Config_Flag(CFG_MULTIPLAYER))
         return;
 
     InitListPolyXlu(play->gs.gfx);
@@ -610,9 +613,9 @@ void comboMultiDrawWisps(GameState_Play* play)
     }
 }
 
-void comboMultiResetWisps(void)
+void Multi_ResetWisps(void)
 {
-    if (!comboConfig(CFG_MULTIPLAYER))
+    if (!Config_Flag(CFG_MULTIPLAYER))
         return;
     bzero(sPlayerWisps, sizeof(sPlayerWisps));
 }