+++ packages/core/src/oot/actors.c
@@ -1,6 +1,10 @@
 #include <combo.h>
 #include <combo/dungeon.h>
 #include <combo/souls.h>
+#include <combo/player.h>
+#include <combo/mask.h>
+#include <combo/config.h>
+#include <combo/global.h>
 
 static s16 sActorIdToSpawn;
 
@@ -264,11 +268,11 @@ Actor* comboSpawnActor(ActorContext* actorCtx, GameState_Play *play, short actor
         variable = 0xff00;
     }
 
-    if (comboConfig(CFG_OOT_OPEN_MASK_SHOP) && play->sceneId == SCE_OOT_MARKET_CHILD_NIGHT && actorId == AC_EN_DOOR)
+    if (Config_Flag(CFG_OOT_OPEN_MASK_SHOP) && play->sceneId == SCE_OOT_MARKET_CHILD_NIGHT && actorId == AC_EN_DOOR)
         if (((variable >> 7 & 7) == 0x5) && ((variable & 0x3f) == 0x10))
             variable = 0x1bf;
 
-    if (comboConfig(CFG_OOT_OPEN_ZD_SHORTCUT) && actorId == AC_BG_SPOT06_OBJECTS && play->sceneId == SCE_OOT_LAKE_HYLIA)
+    if (Config_Flag(CFG_OOT_OPEN_ZD_SHORTCUT) && actorId == AC_BG_SPOT06_OBJECTS && play->sceneId == SCE_OOT_LAKE_HYLIA)
         if (((variable >> 8) & 0xff) == 3)
             return NULL;
 
@@ -284,7 +288,7 @@ static int GetRoomClearFlagForActor(GameState_Play* play, int flag)
     int res;
 
     res = GetRoomClearFlag(play, flag);
-    if (comboConfig(CFG_ER_WALLMASTERS) && sActorIdToSpawn == AC_EN_WALLMAS)
+    if (Config_Flag(CFG_ER_WALLMASTERS) && sActorIdToSpawn == AC_EN_WALLMAS)
         res = 0;
     return res;
 }
@@ -295,7 +299,7 @@ static int shouldActorIgnorePlayer(Actor* this, Actor_Player* link)
 {
     u16 variable;
 
-    if (link->mask != PLAYER_MASK_STONE)
+    if (link->mask != MASK_STONE)
         return 0;
 
     variable = this->variable;